Updates & renamings
