+++ package.json
@@ -15,16 +15,11 @@
     "@types/custom-functions-runtime": "^1.4.0",
     "@types/office-runtime": "^1.0.0",
     "copy-webpack-plugin": "^4.5.2",
-    "file-loader": "^1.1.11",
-    "html-loader": "^0.5.5",
-    "html-webpack-plugin": "^3.2.0",
-    "http-server": "^0.11.1",
     "office-toolbox": "^0.1.0",
     "ts-loader": "^4.4.2",
     "typescript": "^2.7.2",
     "webpack": "^4.17.2",
     "webpack-cli": "^3.1.0",
-    "webpack-command": "^0.4.1",
     "webpack-dev-server": "^3.1.1"
   }
 }
