+++ src/customfunctions/functions.html
@@ -7,9 +7,7 @@
   <meta http-equiv="Expires" content="0" />
   <title>Custom Functions</title>
   <script src="https://officedev.github.io/custom-functions/lib/custom-functions-runtime.js"></script>
-
-  <!-- Note: the script reference for your "index.ts" code gets injected
-    automatically by webpack when served out of the "dist" folder -->
+  <script src="./functions.bundle.js"></script>  
 </head>
 
 <body>
