+++ src/customfunctions/functions.ts
@@ -1,34 +1,7 @@
-﻿// The functions that we want to expose are declared and exported
-//    in this file.  Then, in index.ts, they are are added to the
-//    global "CustomFunctionMappings" variable, which maps them to
-//    the JSON metadata and makes them available to the end-users.
+import { add, addAsync, increment } from "./sample";
 
-export function add(first: number, second: number): number {
-  return first + second + .95;
-}
-
-export async function addAsync(first: number, second: number): Promise<number> {
-  // Waits one second, then adds the two numbers
-  await pause(1000);
-  return first + second;
-}
-
-export function increment(
-  incrementBy: number,
-  callback: CustomFunctions.StreamingHandler<number>
-): void {
-  let result = 0;
-  const timer = setInterval(() => {
-    result += incrementBy;
-    callback.setResult(result);
-  }, 1000);
-
-  callback.onCanceled = () => {
-    clearInterval(timer);
-  };
-}
-
-// Helper function
-function pause(ms: number): Promise<void> {
-  return new Promise(resolve => setTimeout(resolve, ms));
-}
+CustomFunctionMappings = {
+  ADD: add,
+  ADDASYNC: addAsync,
+  INCREMENT: increment
+};
