--- src/customfunctions/index.ts
@@ -1,7 +1,0 @@
-import { add, addAsync, increment } from "./functions";
-
-CustomFunctionMappings = {
-  ADD: add,
-  ADDASYNC: addAsync,
-  INCREMENT: increment
-};
