+++ webpack.config.js
@@ -1,9 +1,8 @@
-const HtmlWebpackPlugin = require("html-webpack-plugin");
 const CopyWebpackPlugin = require("copy-webpack-plugin");
 
 module.exports = {
   entry: {
-    "customfunctions/index": "./src/customfunctions/index.ts"
+    "customfunctions/functions": "./src/customfunctions/functions.ts"
   },
   output: {
     filename: '[name].bundle.js',
@@ -11,7 +10,7 @@ module.exports = {
   },
   devtool: "source-map",
   resolve: {
-    extensions: [".ts", ".tsx", ".html", ".js", "json"]
+    extensions: [".ts", ".js"]
   },
   module: {
     rules: [
@@ -19,27 +18,14 @@ module.exports = {
         test: /\.tsx?$/,
         exclude: /node_modules/,
         use: "ts-loader"
-      },
-      {
-        test: /\.html$/,
-        exclude: /node_modules/,
-        use: "html-loader"
-      },
-      {
-        test: /\.(png|jpg|jpeg|gif)$/,
-        use: "file-loader"
       }
     ]
   },
   plugins: [
-    new HtmlWebpackPlugin({
-      template: "./src/customfunctions/index.html",
-      chunks: ["customfunctions/index"],
-      filename: "customfunctions/index.html"
-    }),
     new CopyWebpackPlugin([
       { from: "./src/customfunctions/metadata", to: "customfunctions/metadata" },
-      { from: "./src/index.html", to: "index.html" },
+      { from: "./src/customfunctions/functions.html", to: "customfunctions/functions.html" },
+      { from: "./src/about.html", to: "about.html" },
     ])
   ],
   devServer: {
